backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0.post1
SQLAlchemy==2.0.23
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
pydantic==1.10.13
email-validator==2.1.0.post1
